Move embed helpers into JSON service
